Reset repository
This project was originally started in August 2015 and then developed on
and off for two months.  Most of the commits from that time were just giant
refactoring patches reversing each others' changes, so the repository
history was re-rooted at a point in time where the project architecture had
mostly stabilized.