packfile/base128: Write{LE,MBE}: return number of bytes written
It appears to be the convention in Go that, when writing and reading, the
number of bytes thus transferred should either be constant, a return value,
or trivially deducible.

Also, prefer arrays to slices for temporary storage: I *think* (suitably
small) arrays are always stack-allocated, whereas slices may not necessarily
be.  (They definitely wouldn't be in the old WriteMBE implementation.)