packfile: Reader: add new Close method for reading and verifying checksum
A convenient idiom for reading an entire packfile is to repeatedly call
Read in a loop on Len() > 0, but this forgets to check the checksum, as
it was only read when one tries to read past the end of the packfile.
Moving this check to a Close method provides a natural way of finishing
the packfile without complicating the loop.

This incidentally made Reader almost symmetrical with Writer, so I
decided to add a Len method to the Writer as well.  Dunno what use it
will ever have, but it looks nice.