packfile: require the presence of a repository when reading
Standalone packfile parsing is nice, but even better is parsing
in constant space - or at the very least, a constant amount of
space per packfile entry (the offset -> ID map) plus a linear
amount of temporary space for the current object.  With this
we have support for thin packfiles also.