protocol: refactor pkt-line scanning
Scanning from pkt-line command streams is a real bitch, as one
has to treat the first line specially (it contains the capability
listing after a null or a space, depending on the command), and
the lines are not required to be newline-terminated, so naive
rune-at-a-time scanning can't know to stop at a pkt-line boundary.
The old scanCmds function tried to abstract all this ugliness away
using reflection, but the pill was perhaps worse than the poison;
its functionality was duplicated in the
protocol.{Upload,Receive}Pack functions.

The special case of the first line is succintly handled by simply
trying to scan the capability listing from every line.  This will
obviously error out from the second line onwards, but the rest of
the line will still have been parsed correctly, so one need only
care about the error if it is returned halfway into the line.
(There exists now an obvious error case where false capability
listings in the middle of a pkt-line stream will overwrite the
"real" capability listing rather than return an error, but
reprimanding misbehaving clients was not a priority of this
implementation.)

Other minor refactorings in the protocol code are also included.