pktline: move to a line-based interface
pktline.Reader and -Writer were originally implemented as io.Readers
and -Writers so that one could directly fmt.Fscanf from and -Fprintf
to them.  Though this was a nice API, it was also kind of fragile: with
fmt.Fprintf you had to assume that it would print its formatted output
in a single write, and with fmt.Fscanf that pkt-line boundaries would
also be marked within the payloads themselves.  Rather than rely on
these assumptions, or crudely work around them, I opted to redesign
the interface such that a whole pkt-line is now the basic unit of
interaction.  Formatted scanning and printing is now done through the
line{Reader,Writer} interfaces and the fmt{Lscanf,Lprintf} functions
in the protocol package.  (The only other use of formatted pkt-line
printing was in package protocol/http, which was easy enough to
hand-inline.)

The package pktline internals were also completely reworked.  Using
fmt.Fscanf to get the length of a pkt-line was overkill, in that
not only did it pull in reflection, scanning four bytes also requires
four one-byte Reads from an unbuffered reader.  Better to just fetch
them in one Read and use a hand-written decoder to parse the hexadecimal
length.  For symmetry, writing the length was changes to operate
similarly.