protocol: UploadPack: implement packing heuristics
writePack now sorts its objects first by type and then by size before
writing them out into the packfile.  This is almost the same as the
heuristics used by the reference implementation [1], with the important
exception that blobs and trees aren't sorted by their base names.
This is unfortunate, as files with the same name are likely to delta
extremely well against each other, but I wasn't willing to implement
the preprocessing and bookkeeping data necessary for this.

Sorting by type is not strictly necessary in our implementation, as
packfile.Writer maintains a separate delta window (of one object)
for each object type; but it helps avoid pointless size comparisons,
and may come in useful if the implementation ever changes.

[1]: https://www.kernel.org/pub/software/scm/git/docs/technical/pack-heuristics.html