object: Tree: use pointer receivers for methods
This is perhaps a bit silly, but I find value in the Interface
implementors consistently being pointer types.  This additionally
makes it possible to Unmarshal data into a zero Tree value.  This
makes iterating over or indexing a Tree value returned by the package
functions slightly less convenient, but the same problem already
existed for the Blob type to no great frustration.

Also, add and change comments to highlight how the Tree type is
semantically different from the others.