repository: refactor Interface
I could not find a nice way to reconcile storing two different kinds
of objects (refs and symrefs) in the same namespace in the interface,
so I simply chose not to and cut symrefs out entirely.  This isn't
even too limiting, as according to the Git developers [1], the only
valid types of symrefs are HEAD and refs/remotes/<remote>/HEAD,
the latter of which is out of scope for the interface anyway.
So, the symref methods could be replaced with {Get,Set}HEAD.

UpdateRef was also inlined as a repository method in order to
enforce implementation-level atomicity guarantees for it.

The Init method was removed from the interface, as it should
properly be the job of a function.

While updating repository/appengine to match the new interface,
some internal refactorings were also performed.

[1]: https://stackoverflow.com/questions/4986000/whats-the-recommended-usage-of-a-git-symbolic-reference#answers