object: Marshal: new return value: ID
An object's ID is computed from its binary representation, so it
is useful to provide an interface that computes both, rather than
require users to marshal an object twice (once in MarshalBinary
and once in Hash) for this use case.

repository/appengine was patched to use this implementation.
packfile didn't need it.