pktline: Reader.ReadLine: fix and document error returns
The documentation says that the error is io.ErrUnexpectedEOF if
an EOF is encountered in the middle of a pkt-line.  This should
hold even if the "middle" is right before or after a pkt-line
length, but io.ReadFull returns io.EOF in these cases (because
n == 0).  Catch them and return io.ErrUnexpectedEOF instead.

This behavior does have the problem that a clean end of file
immediately after a flush-pkt cannot be signaled unambiguously,
but I did not wish to invent a separate EOF value for the package
or have io.EOF pull double duty, so the best we can do at the
moment is document it.