protocol: bugfixes and refactoring
protocol: AdvertiseRefs: fix capability declarations

AdvertiseRefs forgot to list the server capabilities in the event
that HEAD was not a valid ref, but other valid refs existed.  In
the new code, HEAD is now also peeled if it points at an annotated
tag, though I do not know if such a ref is a legal value for HEAD.

The canonical implementation's ls-remote and show-ref commands omit
refs that point at nonexistent objects, and print errors for them.
The protocol specification does not seem to require this, but I
documented our deviation from it just in case.

protocol: Capabilities: list in alphabetical order

protocol: ReceivePack, UploadPack: return early if no commands are received

Such requests are invalid per the spec, but this simplifies
reimplementing the canonical receive-pack and upload-pack commands,
where a flush-pkt after the reference discovery part (AdvertiseRefs)
ends the process cleanly.

protocol: negotiate: handle end-of-stream

In the HTTP protocol packfile negotiation may take place over two
POST requests, the first of which ends without a "done\n", so negotiate
must be able to terminate gracefully on end-of-body.  Because
pktline.Reader.ReadLine overloads this meaning atop io.ErrUnexpectedEOF,
comments were added to both EOF-like cases to clarify their meanings.

While there, the order of operands in the io.EOF case if statement
was switched, because I think it's clearer if multiAck comes first.