repository: refactor error values
The old two error variables weren't enough to cover all possible
error cases in Interface.UpdateRef.  The two new ones, ErrRefMismatch
and ErrRefExist, are returned when the ref's current value does not
match the expected one and when the ref already exists respectively.
ErrNotExist was also split into ErrRefNotExist and ErrObjectNotExist
in order to be able to distinguish between the cases where the named
ref does not exist and where newID points to a nonexistent object.

With the ref/object not exist error split, the mapErr function in
repository/appengine was moved to refs.go and renamed mapRefErr,
where it now only handles mapping datastore.ErrNoSuchEntity to
repository.ErrRefNotExist.  In practice it did this already, as
the object methods have their own error handling, but this makes
the intent more explicit.