protocol: UploadPack: fix multi_ack_detailed implementation
Ironically this implementation has more BUGs than the old one,
but it actually does what the capability advertises: explicit
reporting when the server has figured out a minimal packfile.
This also permitted us to add support for the no-done capability,
so HTTP upload-pack transactions can now be completed in a single
POST request.